moved di-container into a separate plugin and repository
