--- src/main/java/core/di/contract/JavaInstanceConfigurationInterface.java
@@ -1,7 +1,0 @@
-package core.di.contract;
-
-import java.util.LinkedHashMap;
-
-public interface JavaInstanceConfigurationInterface {
-    LinkedHashMap<Class<?>, Class<?>[]> getInstanceConfigure();
-}
