--- src/main/java/core/di/exception/NoSuchInstanceException.java
@@ -1,7 +1,0 @@
-package core.di.exception;
-
-public class NoSuchInstanceException extends RuntimeException {
-    public <T> NoSuchInstanceException(Class<T> clazz) {
-        super("No such instance '%s'".formatted(clazz.getName()));
-    }
-}
